Restructure repo, build batch img processing
