+++ papers/meso/archive/diagrams.txt
