+++ papers/meso/archive/output.txt
