+++ papers/meso/data/img/stad1.1.png
