+++ papers/meso/data/img/stad1.2.png
