+++ papers/meso/data/img/stad2.1.png
