+++ papers/meso/data/img/stad2.2.png
